% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_utilities.R
\name{change_defaults}
\alias{change_defaults}
\title{Change Method Defaults}
\usage{
change_defaults(method, ...)
}
\arguments{
\item{method}{Function definition.}

\item{...}{New default values for the arguments}
}
\description{
Change Method Defaults
}
