% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_utilities.R
\name{clean_method_environment}
\alias{clean_method_environment}
\title{Clean Method Environment}
\usage{
clean_method_environment(e)
}
\arguments{
\item{e}{Environment containing methods in an object}
}
\value{
There is no return value. The function is called for
its side-effect of cleaning a method environment.
}
\description{
Clean the environment of a method (or methods) so that
they contain a single object -- self -- which is the
environment defining an object
}
