% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{Implementation}
\alias{Implementation}
\title{Implementation Class}
\usage{
Implementation()
}
\description{
Initialize an object with concrete implementations of abstract method
definitions.
}
\details{
\link[=inheritance]{Inherit}
from \code{Implementation} (using the
\code{\link{implements}} utility function) if you want
your class to implement an \code{\link{Interface}}.
}
\examples{
BinaryOperation = function() {
  self = Interface()
  self$operate = function(x = numeric(1L), y = numeric(1L)) return(numeric(1L))
  return_object(self, "BinaryOperation")
}
Add = function() {
  self = implements(BinaryOperation)
  self$operate = function(x = numeric(1L), y = numeric(1L)) return(x + y)
  return_object(self, "Add")
}
Multiply = function() {
  self = implements(BinaryOperation)
  self$operate = function(x = numeric(1L), y = numeric(1L)) return(x * y)
  return_object(self, "Multiply")
}
Add()$operate(1, 1)
Multiply()$operate(2, 2)

}
