% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{MappedAllTest}
\alias{MappedAllTest}
\title{Mapped All Test}
\usage{
MappedAllTest(basic_tester)
}
\arguments{
\item{basic_tester}{An object that can be converted to a function}
}
\description{
Test that all \code{\link{MappedTest}} results are \code{\link{TRUE}}
}
