% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{MappedAnyTest}
\alias{MappedAnyTest}
\title{Mapped Any Test}
\usage{
MappedAnyTest(basic_tester)
}
\arguments{
\item{basic_tester}{An object that can be converted to a function}
}
\description{
Test that any \code{\link{MappedTest}} results are \code{\link{TRUE}}
}
