% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{MappedSummarizer}
\alias{MappedSummarizer}
\title{Mapped Summarizer}
\usage{
MappedSummarizer(...)
}
\arguments{
\item{...}{A list of summarizing functions.}
}
\value{
Object of class \code{\link{TestSteps}} that summarizes each
element of objects to test.
}
\description{
Apply a \code{\link{Summarizer}} to each element of a list, in order
to test that a particular summary of each lists item meets a certain
criterion. \code{MappedSummarizer}s are typically included in
\code{\link{TestPipeline}}s.
}
