% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{MappedTest}
\alias{MappedTest}
\title{Mapped Test}
\usage{
MappedTest(basic_tester, boolean_aggregator)
}
\arguments{
\item{basic_tester}{An object that can be converted to a function}

\item{boolean_aggregator}{A function that summarizes a \code{\link{logical}}
vector.}
}
\description{
Apply a \code{\link{Test}} to each element of a list.
}
