% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_apply.R
\name{method_apply}
\alias{method_apply}
\title{Method Apply}
\usage{
method_apply(objects, method_name, ...)
}
\arguments{
\item{objects}{List of objects.}

\item{method_name}{Character string giving the name of the method.}

\item{...}{Arguments to pass to the method.}
}
\description{
Call a method for each item in a list of objects.
}
