% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{MultiTest}
\alias{MultiTest}
\alias{All}
\alias{Any}
\title{Multi Test}
\usage{
MultiTest(test_function_list, boolean_aggregator)

All(...)

Any(...)
}
\arguments{
\item{test_function_list}{List of objects of class \code{\link{TestSteps}}
or \code{\link{function}}.}

\item{boolean_aggregator}{A function that summarizes a \code{\link{logical}}
vector.}
}
\value{
Object of class \code{\link{TestSteps}} that tests several
criteria at the same time.
}
\description{
Assess several criteria.
}
\section{Functions}{
\itemize{
\item \code{All()}: Test that all of the criteria are met.

\item \code{Any()}: Test that any of the criteria are met.

}}
\examples{
is_matrix = All(
 is.numeric,
 TestPipeline(
   Summarizer(dim, length),
   TestRange(0, 2)
 )
)
is_matrix$apply(array("a", c(1))) # FALSE
is_matrix$apply(array("a", c(1, 1, 2))) # FALSE
is_matrix$apply(array(1, c(1, 1, 2))) # FALSE
is_matrix$apply(array(1, c(1, 2))) # TRUE
is_matrix$apply(1) # TRUE

}
