% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{Not}
\alias{Not}
\title{Not}
\usage{
Not(basic_tester)
}
\arguments{
\item{basic_tester}{An object that can be converted to a function}
}
\value{
Object of class \code{\link{Test}} that evaluates the
complement of \code{basic_tester}.
}
\description{
Not
}
\examples{
Not(is.numeric)$apply(1) # FALSE
Not(is.numeric)$apply("1") # TRUE

}
