% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{oor_debug}
\alias{oor_debug}
\title{Debug Methods}
\format{
This list is an object of class \code{DebugList} with the following
methods.
\itemize{
\item \verb{$add(method, class)} -- Add a \code{method} in a particular \code{class} to be
debugged. By default \code{class == "Base"}, which will typically debug a method
called \code{method} in any class because typically all classes will inherit
from \code{Base}.
\item \verb{$remove(method)} -- Stop debugging all \code{method}s with a particular name.
\item \verb{$list()} -- List all \code{method}-\code{class} pairs that would be debugged if
called.
\item \verb{$undebug(method)} -- Stop debugging a method in all objects.
}
}
\usage{
oor_debug
}
\description{
List of methods that will be flagged for debugging when they are created.
}
\details{
Because oor methods are not created until the object within which they
exist is created, users cannot as easily use the \code{\link{debug}}
function. By adding method-class pairs to this \code{oor_debug} object,
users can control what methods get debugged before they are created.
Note that once a method that is flagged for debugging is created, it cannot
be \code{\link{undebug}}ed. To stop debugging particular methods, the
associated objects must be recreated and this is often best done by just
reruning the script.
}
\keyword{datasets}
