% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_utilities.R
\name{return_facade}
\alias{return_facade}
\title{Return Facade}
\usage{
return_facade(self, private, class)
}
\arguments{
\item{self}{New object.}

\item{private}{Environment to use for containing private methods and fields.}

\item{class}{String giving the class name.}
}
\description{
Experimental
}
