% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_utilities.R
\name{return_object}
\alias{return_object}
\title{Return Object}
\usage{
return_object(self, class)
}
\arguments{
\item{self}{New object.}

\item{class}{String giving the class name.}
}
\value{
New object of class given by \code{class}.
}
\description{
This should be the final function called in a class
definition. Think of it like return(...)
}
