% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_utilities.R
\name{save_parent_methods}
\alias{save_parent_methods}
\title{Save Parent Methods}
\usage{
save_parent_methods(object, parent_name = "parent")
}
\arguments{
\item{object}{Object for saving the methods of that object in a list
called \code{parent} (or whatever is passed to \code{parent_name}).}

\item{parent_name}{Name of the list to be placed in \code{object} containing
references to the methods in \code{object}.}
}
\description{
Save the methods in an object in a list so that they can be used
in methods that will be overridden.
}
\examples{
A = function(x) {
  self = Base()
  self$x = x
  self$f = function() self$x
  return_object(self, "A")
}
B = function(x) {
  self = A(x)
  save_parent_methods(self)
  self$f = function() self$parent$f() + 100
  return_object(self, "B")
}
A(100)$f() ## 100
B(100)$f() ## 200

}
