% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{Summarizer}
\alias{Summarizer}
\title{Summarizer}
\usage{
Summarizer(...)
}
\arguments{
\item{...}{A list of summarizing functions.}
}
\value{
Object of class \code{\link{Test}} that summarizes objects
to test.
}
\description{
Summarize an object to be tested, so that the test is applied to the
summary and not the object itself (e.g. \code{length(dim(object)) == 2L}).
\code{Summarizer}s are typically included in \code{\link{TestPipeline}}s.
}
