% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{Testable}
\alias{Testable}
\title{Testable Class}
\usage{
Testable()
}
\description{
Initialize an object with functionality for validating objects
}
\details{
\link[=inheritance]{Inherit} from \code{Testable} if you
would like your class to provide a validity
check. Validity checking often requires differentiating between public
versus private members, as well as fields versus methods.
}
\examples{

Printer = function(x) {
  self = Testable()
  self$.x = x
  self$valid = function() {
     if (!is.character(self$.x)) {
       return("can only print character strings")
     }
     if (length(self$.x) != 1L) {
       return("can only print length-1 character vectors")
     }
     return(TRUE)
  }
  self$print = function() print(self$.x)
  return_object(self, "Printer")
}
printer = Printer("something to print")
printer$print()
try(Printer(0)) ## error

}
