% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestBasic}
\alias{TestBasic}
\title{Basic Test}
\usage{
TestBasic(basic_tester)
}
\arguments{
\item{basic_tester}{An object that can be converted to a function}
}
\value{
Object of class \code{\link{Test}} that evaluates
the \code{basic_tester}.
}
\description{
Basic Test
}
