% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestHomo}
\alias{TestHomo}
\title{Test for Homogeneity}
\usage{
TestHomo()
}
\value{
Object of class \code{\link{TestSteps}} that tests that all
elements in an object are identical.
}
\description{
Test that all elements in an object are identical.
}
