% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestLenZero}
\alias{TestLenZero}
\title{Length-Zero Vector Test}
\usage{
TestLenZero()
}
\value{
Object of class \code{\link{Test}} that tests that a vector
has zero length.
}
\description{
Test that the input is a vector of length zero, and prints the
elements that exist.
}
