% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestPipeline}
\alias{TestPipeline}
\title{Test Pipeline}
\usage{
TestPipeline(...)
}
\arguments{
\item{...}{Objects of class \code{\link{TestSteps}} or \code{\link{function}}.
The final object in this list must be either a \code{TestSteps} object
with an \code{apply} method that returns a length-one logical vector
or a function that does so.}
}
\value{
Object inheriting from \code{\link{TestSteps}}
}
\description{
Test Pipeline
}
\examples{
is_matrix = TestPipeline(
  Summarizer(dim, length),
  TestRange(0, 2)
)
is_matrix$apply(array("a", c(1))) # TRUE
is_matrix$apply(array(1, c(1, 2, 3))) # FALSE

each_is_matrix = TestPipeline(
  MappedSummarizer(dim, length),
  All(TestRange(0, 2))
)
each_is_matrix$apply(list(1, matrix(1, 2, 3), "a")) # TRUE
each_is_matrix$apply(list(1, array(1, c(2, 3, 4), "a"))) # FALSE

}
