% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestRange}
\alias{TestRange}
\title{Range Test}
\usage{
TestRange(lower, upper)
}
\arguments{
\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\value{
Object of class \code{\link{TestSteps}} that tests that all
elements in an object numerically on a particular range.
}
\description{
Test that all elements in an object greater than or equal to \code{lower}
and less than or equal to \code{upper}.
}
