% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestSteps}
\alias{TestSteps}
\alias{as.function.TestSteps}
\title{Abstract Class Testing Objects}
\usage{
TestSteps()

\method{as.function}{TestSteps}(x, ...)
}
\value{
Object with an \code{apply} method that takes a single
argument, \code{x}, and returns a length-one \code{\link{logical}} vector.
}
\description{
Abstract Class Testing Objects
}
