% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classes.R
\name{TestSubset}
\alias{TestSubset}
\title{Subset Test}
\usage{
TestSubset(set)
}
\arguments{
\item{set}{Universe of possibilities.}
}
\value{
Object of class \code{\link{Test}} that tests that all
elements in an object are in a particular set.
}
\description{
Test that all elements in an object are in \code{set}
}
