% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_utilities.R
\name{validate_object}
\alias{validate_object}
\title{Validate Object}
\usage{
validate_object(object)
}
\arguments{
\item{object}{Object to be validated.}
}
\value{
TODO -- check $valid methods
}
\description{
S3 generic for checking the validity of a constructed
object. should either return nothing or trigger an error.
Typically these S3 methods will make use of object$valid
methods.
}
