% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{Base}
\alias{Base}
\title{Base Class}
\usage{
Base()
}
\description{
Initialize an empty object.
}
\details{
\link[=inheritance]{Inherit} from \code{Base} if you want to start
from an empty class
}
\examples{
empty_object = Base()
print(empty_object)
names(empty_object)

Printer = function(x) {
  self = Base()
  self$.x = x
  self$print = function() print(self$.x)
  return_object(self, "Printer")
}
printer = Printer("something to print")
printer$print()

}
