% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_classes.R
\name{ValidityMessager}
\alias{ValidityMessager}
\title{Validity Messager}
\usage{
ValidityMessager(test_function, fail_message)
}
\arguments{
\item{test_function}{Object that is coercible to a \code{\link{function}},
typically a function that will return a length-1 \code{\link{logical}}
vector. Often this object will inherit from \code{\link{TestSteps}},
which provides a way to compose object tests.}

\item{fail_message}{Length-1 \code{\link{character}} vector to display
if \code{test_function} does not return \code{TRUE}.}
}
\value{
Object of class \code{ValidityMessager} containing a \code{check}
method that will return \code{TRUE} or fail with \code{fail_message}.
}
\description{
Couple a test function and failure message, so that if the test function
does not return \code{TRUE} then the failure message is given.
}
\details{
This class useful when building \code{valid} methods in
\code{\link{Testable}} classes.
}
\examples{
ValidityMessager(is.numeric, "not numeric")$check("1") ## "not numeric"

HoldANumber = function(x) {
  self = Testable()
  self$x = x
  self$.numeric_validity = ValidityMessager(is.numeric, "not numeric")
  self$valid = function() {
    self$.numeric_validity$check(self$x)
  }
  return_object(self, "G")
}
try(HoldANumber("a")) ## error message
HoldANumber(1) ## success

}
